Rewrite mark_stack_map from scratch. Passes all tests :)
Finally, I'm actually confident that this is doing the right thing!

Key insight was to use the structure of the dead sections and mark
"everything else". Normally in marking, we use the structure of the
live data, so this is a reversal. But now seems obvious that it's
the right way.

This is because the completed calls are so easy to identify in the
structure of the heap. A stack-map Pop links directly back to a
stack-map Push, and everything in between is dead.

By contrast, heap sections for live calls are complicated. They're
perforated with dead parts corresponding to deeper calls. And
only the most recent call at each depth level is live. Messy.

Note:
Also need to just do a normal trace from everything marked.