Fix kernel function evaluation
It was applying one value at a time which is _usually_ OK.
But some kernel functions are not curried, need 2+ args.
Need to calculate how many args to apply.
Edge case: Time.here has length=0 but gets Unit applied to it.
This logic always applies as many args as possible, but
a simpler version could just do `kernelFn.length || 1`.