Rename things
This experiment is not really about bytecode, it's about using an explicit stack
In fact my conclusion is that explicit stack is probably slow because it can't use
machine registers.
The point here was to avoid the replay system but the replay system allows us to
use registers. So the bodies of functions should be optimally fast.
Replay version does allocate the stack map and that irks me a bit.
But the explicit stack does so many memory stores and loads, isn't that worse?