Special-case Process struct to avoid JS infinite loop.
JS wrapper was trying to evaluate the full Process tree,
but it's an infinite loop by design!

Luckily, it also has a well-defined lifetime and a unique ID,
which is what we need to manage cross-language references.

When reading a Process out to JS we only decode the ID.
When writing a Process back to Wasm, we look it up by ID in
an array to convert back to an address.
This indirection is needed because the address could have
changed in the meantime if there was a major GC.

Also used this as an excuse to make a DynamicArray, which
I've often thought about. Should be handy in other places.

I haven't yet tested the process cache invalidation, so
I could be wrong. For now, don't actually remove anything,
just print out debug messages.