Fix compactor bug
We were only adjusting pointer addresses while moving the parent.
But it's possible for a child to move while the parent doesn't, and
that case was broken.

If the parent isn't moving, then it must be in the first live patch,
so we now iterate over that as a special case before the main loop.