Resolve ambiguity between Records and JS Kernel objects on Wasm write
If we don't recognise the FieldGroup, treat the value as a JsRef.

If it turns out to be a Kernel object, it will just come out of Wasm again.
If it turns out to be a Record, then Wasm code will either call
Utils_access or Utils_update on it. At that point we can detect the JsRef
and convert it, knowing for sure it's a Record.

This neat little trick fixes a thorny issue that kept coming up,
especially for VirtualDom and related packages like Markdown and WebGl.

Incurs a tag check on every Record access & update but that's cheap enough